refactored and added a student quiz page
