--- frontend/pages/student.tsx
@@ -1,10 +1,0 @@
-import StudentView from "../components/StudentView/StudentView";
-
-  
-  const Student = props => (
-        <>
-            <StudentView />
-        </>
-  );
-  
-  export default Student;
